2010-04-25	Joel Sherrill <joel.sherrilL@OARcorp.com>
	* libchip/rtc/ds1375-rtc.h, libchip/rtc/ds1375.c,
	libchip/rtc/icm7170.c, libchip/rtc/icm7170_reg.c,
	libchip/rtc/m48t08_reg.c, libchip/rtc/rtc.h: Use uintptr_t instead of
	uint32_t.

+++ c/src/libchip/rtc/ds1375-rtc.h
@@ -64,10 +64,10 @@ bool
 rtc_ds1375_device_probe( int minor );
 
 uint32_t
-rtc_ds1375_get_register( uint32_t port, uint8_t reg );
+rtc_ds1375_get_register( uintptr_t port, uint8_t reg );
 
 void
-rtc_ds1375_set_register( uint32_t port, uint8_t reg, uint32_t value );
+rtc_ds1375_set_register( uintptr_t port, uint8_t reg, uint32_t value );
 
 /*
  * BSP must supply string constant argument 'i2cname' which matches
@@ -84,11 +84,11 @@ rtc_ds1375_set_register( uint32_t port, uint8_t reg, uint32_t value );
 #define DS1375_RTC_TBL_ENTRY(i2cname) \
 {                                              	      \
 	sDeviceName:	"/dev/rtc",                       \
-	deviceType:		RTC_CUSTOM,                       \
-	pDeviceFns:		&rtc_ds1375_fns,                  \
+	deviceType:	RTC_CUSTOM,                       \
+	pDeviceFns:	&rtc_ds1375_fns,                  \
 	deviceProbe:	rtc_ds1375_device_probe,          \
-	ulCtrlPort1:	(uint32_t)(i2cname),              \
-	ulDataPort:		0,                                \
+	ulCtrlPort1:	(uintptr_t)(i2cname),             \
+	ulDataPort:	0,                                \
 	getRegister:	rtc_ds1375_get_register,          \
 	setRegister:	rtc_ds1375_set_register,          \
 }
